chore(branching): restore original order and tidy up
In splitting branching into two episodes (based on feedback) I mistakenly moved "Git Hygiene" to be the first
chapter. This meant that the first two tasks for which uses issue templates and the state of the repository
`python-maths` was incorrect. I've therefore moved "Git Hygiene" to be the third episode. as I've insufficient time to
reconfigure everything at the moment.

Reverted that and tweaked the first branching episode.