refactor: move conda env to hook episode
Closes #146

Moves setting up Conda/Virtual envs to the hooks episode (which may never get taught in reality!).

This lowers the amount of setup required as we don't really need virtual environment other than to install pre-commit
and run it. The material is left should people wish to go and learn about it.